Add User..Controller,Repository, Tach request createProduct va UploadProductImages (VD 20->24)

+++ shopapp-be/shopapp/src/main/java/com/project/shopapp/dtos/UserDTO.java
@@ -28,9 +28,9 @@ public class UserDTO {
     @JsonProperty("date_of_birth")
     private Date dateOfBirth;
     @JsonProperty("facebook_id")
-    private int facebookId;
+    private Long facebookId;
     @JsonProperty("google_id")
-    private int googleId;
+    private Long googleId;
     @NotNull(message = "RoleId is required")
     @JsonProperty("role_id")
     private Long roleId;
